Updated the lexing regex, so decimal fractions are lexed as a whole token, unseparated from the floating point

+++ src/parser.py
@@ -369,7 +369,7 @@ def Value():
     ntype = NodeType.Value
 
     if CheckToken(token, TokenType.VALUE):
-        val = int(token.get())
+        val = float(token.get())
         token.next()
     elif CheckToken(token, TokenType.IDENT):
         ntype = NodeType.FunctionCall
